% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paft_pkg.R
\docType{package}
\name{paft-package}
\alias{paft-package}
\alias{_PACKAGE}
\alias{paft-packages}
\title{paft: Penalized accelerated failure model}
\description{
The package offers
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/stc04003/paft}
  \item Report bugs at \url{http://github.com/stc04003/paft/issues}
}

}
\author{
\strong{Maintainer}: Ying Chen \email{yxc170430@utdallas.edu}

Authors:
\itemize{
  \item Chuan-Fa Tang \email{Chuan-Fa.Tang@utdallas.edu}
  \item Sy Han Chiou \email{schiou@utdallas.edu}
}

}
