% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paftgee.R
\name{paft.control}
\alias{paft.control}
\title{Control list}
\usage{
paft.control(
  tol = 0.001,
  maxit = 500,
  init = NULL,
  nfold = 0,
  cv.rule = c("cv", "1se", "bic"),
  lambda.vec = 1:100/50,
  pindex = NULL,
  parallel = FALSE,
  parCl = NULL
)
}
\description{
Control list
}
