% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paftgee.R
\name{paftgee}
\alias{paftgee}
\title{Function to fit...}
\usage{
paftgee(
  formula,
  data,
  subset,
  id = NULL,
  weights = NULL,
  lambda = 0,
  penalty = c("SCAD", "Lasso", "aLasso"),
  B = 0,
  corstr = c("independence", "exchangeable", "AR1"),
  control = list()
)
}
\arguments{
\item{formula}{a formula object}

\item{data}{an optional data frame}

\item{subset}{an optional logical vector}

\item{id}{an optional}

\item{weights}{an optional}

\item{lambda}{a numeric value}

\item{B}{a numeric value}

\item{control}{a list}
}
\description{
Function to fit...
}
